requirements.txt
flask
flask-restful
Jinja2
pytorch
numpy
matplotlib
opencv
